ml-service/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0

# Data & ML Core
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2

# Deep Learning
tensorflow==2.15.0

# NLP / Sentiment
transformers==4.36.0
torch==2.1.2
sentencepiece==0.1.99

# Technical Analysis
ta==0.11.0
yfinance==0.2.33

# Database
pymongo==4.6.1

# Utilities
requests==2.31.0
